requirements.txt
# ===== WEB FRAMEWORK =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===== DOCUMENT GENERATION =====
python-docx==1.1.0
lxml==4.9.3

# ===== NLP & MACHINE LEARNING =====
# UDPipe - Монгол хэлний морфологи шинжилгээ
ufal.udpipe==1.3.0.1

# Ollama - Локал LLM (ЧУХАЛ!)
ollama==0.1.6

# ===== TEXT PROCESSING =====
regex==2023.10.3

# ===== UTILITIES =====
python-dateutil==2.8.2
pydantic==2.5.0

# ===== OPTIONAL: Transformers (хэрэв mT5 ашиглах бол) =====
# Одоогоор ашиглахгүй байна, Ollama ашиглана
# transformers==4.35.2
# torch==2.1.1
# sentencepiece==0.1.99

# ===== DEVELOPMENT (Optional) =====
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0